+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 - First Commit. 
 (Bruno Mendoza) [#]()
 ### 1.0.0 First Documentation 25-07-2024
+Feature:
 -To describe architecture, well practices, gitflow, ChangeLog use and proyect information
 (Bruno Mendoza) [#1](https://github.com/DR-Mexico/dr.intranet.web/pull/1)
+
+### 1.1.0 First Test 25-07-2024
+Feature:
+-To get all components tested with vitest
+(Bruno Mendoza) [#2](https://github.com/DR-Mexico/dr.intranet.web/pull/2)